Split Code into files based on Type Rename Print Methods to String and GoString to support fmt.Printf
